% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorages.R
\docType{data}
\name{anchorages_all}
\alias{anchorages_all}
\title{Anchorages data from Global Fishing Watch}
\format{
\subsection{\code{anchorages_all} A data frame with 166,496 rows and 2 columns:}{

\describe{
\item{x}{x coordinate}
\item{y}{y coordinate}
}
}
}
\source{
\url{https://globalfishingwatch.org/datasets-and-code-anchorages/}
}
\usage{
anchorages_all
}
\description{
Data is from Global Fishing Watch and identifies anchorages as anywhere
vessels with AIS remain stationary for 12 hours or more.
}
\details{
Intended for use by \code{oceandatr::get_dist()} with argument \code{dist_to = "anchorages_all"}.

See \code{data-raw/gfw_anchorages.R} for data preparation code.
}
\keyword{datasets}
