% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorages.R
\docType{data}
\name{anchorages_grouped}
\alias{anchorages_grouped}
\title{Grouped anchorages data from Global Fishing Watch}
\format{
\subsection{\code{anchorages_grouped} A data frame with 166,496 rows and 2 columns:}{

\describe{
\item{on_land}{\code{logical} specifying if point is on land (TRUE) or not (FALSE)}
\item{x}{x coordinate}
\item{y}{y coordinate}
}
}
}
\source{
\url{https://globalfishingwatch.org/datasets-and-code-anchorages/}
}
\usage{
anchorages_grouped
}
\description{
Data is from Global Fishing Watch and identifies anchorages as anywhere
vessels with AIS remain stationary for 12 hours or more.  Anchorages close
together have the same names, so to reduce the number of anchorages, they
were aggregated by iso3 code (country code) and label (name) and the mean
longitude and latitude coordinates obtained to get one anchorage point per
name in each country.
}
\details{
Anchorages that fall on land (\code{on_land} column value is TRUE) were defined by
points that fall within the Natural Earth land boundaries, buffered by 10km
inland so as to avoid cutting off coastal anchorages that fall within the
land boundary, due to inaccuracies in the Natural Earth land boundaries, e.g.
for islands and other small scale coastlines. This further reduces the number
of anchorages by removing any that are along rivers and so not relevant to
most ocean based data work.

Intended for use by \code{oceandatr::get_dist()} with argument \code{dist_to = "anchorages_grouped"} or \code{dist_to = "anchorages_land_masked"}.

See \code{data-raw/gfw_anchorages.R} for data preparation code.
}
\keyword{datasets}
